First step splitting crates from api; tiledb-common passes tests on all features
