+++ Cargo.toml
@@ -2,8 +2,11 @@
 resolver = "2"
 members = [
   "tiledb/api",
+  "tiledb/common",
   "tiledb/proc-macro",
+  "tiledb/proptest-integrations",
   "tiledb/queries",
+  "tiledb/serde",
   "tiledb/sys",
   "tiledb/test-utils",
   "tiledb/utils",
@@ -11,8 +14,11 @@ members = [
 ]
 default-members = [
   "tiledb/api",
+  "tiledb/common",
   "tiledb/proc-macro",
+  "tiledb/proptest-integrations",
   "tiledb/queries",
+  "tiledb/serde",
   "tiledb/sys",
   "tiledb/utils",
 ]
@@ -24,10 +30,18 @@ version = "0.1.0"
 
 [workspace.dependencies]
 anyhow = "1.0"
+arrow = { version = "52.0.0", features = ["prettyprint"] }
+arrow-schema = { version = "52.0.0" }
+num-traits = "0.2"
+paste = "1.0"
 proptest = { version = "1.0.0" }
-serde_json = { version = "1.0.114", features = ["float_roundtrip"] }
-tiledb = { path = "tiledb/api", version = "0.1.0" }
+serde = { version = "1", features = ["derive"] }
+serde_json = { version = "1", features = ["float_roundtrip"] }
+thiserror = { version = "1" }
+tiledb-api = { path = "tiledb/api", version = "0.1.0" }
+tiledb-common = { path = "tiledb/common", version = "0.1.0" }
 tiledb-proc-macro = { path = "tiledb/proc-macro", version = "0.1.0" }
+tiledb-serde = { path = "tiledb/serde", version = "0.1.0" }
 tiledb-sys = { path = "tiledb/sys", version = "0.1.0" }
 tiledb-test-utils = { path = "tiledb/test-utils", version = "0.1.0" }
 tiledb-utils = { path = "tiledb/utils", version = "0.1.0" }
