+++ tiledb/api/src/array/dimension/mod.rs
@@ -1,17 +1,15 @@
 use std::fmt::{Debug, Display, Formatter, Result as FmtResult};
 use std::ops::Deref;
 
-use serde::{Deserialize, Serialize};
-use serde_json::json;
-use util::option::OptionSubset;
-
 use crate::array::CellValNum;
 use crate::context::{CApiInterface, Context, ContextBound};
 use crate::datatype::PhysicalType;
 use crate::error::{DatatypeErrorKind, Error};
 use crate::filter::list::{FilterList, FilterListData, RawFilterList};
 use crate::range::SingleValueRange;
-use crate::{physical_type_go, Datatype, Factory, Result as TileDBResult};
+use crate::{physical_type_go, Datatype, Result as TileDBResult};
+
+pub use tiledb_common::array::dimension::DimensionConstraints;
 
 pub(crate) enum RawDimension {
     Owned(*mut ffi::tiledb_dimension_t),
@@ -143,17 +141,6 @@ impl Dimension {
     }
 }
 
-impl Debug for Dimension {
-    fn fmt(&self, f: &mut Formatter) -> FmtResult {
-        let data =
-            DimensionData::try_from(self).map_err(|_| std::fmt::Error)?;
-        let mut json = json!(data);
-        json["raw"] = json!(format!("{:p}", *self.raw));
-
-        write!(f, "{}", json)
-    }
-}
-
 impl PartialEq<Dimension> for Dimension {
     fn eq(&self, other: &Dimension) -> bool {
         eq_helper!(self.name(), other.name());
@@ -167,278 +154,6 @@ impl PartialEq<Dimension> for Dimension {
         });
 
         true
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, OptionSubset, PartialEq, Serialize)]
-pub enum DimensionConstraints {
-    Int8([i8; 2], Option<i8>),
-    Int16([i16; 2], Option<i16>),
-    Int32([i32; 2], Option<i32>),
-    Int64([i64; 2], Option<i64>),
-    UInt8([u8; 2], Option<u8>),
-    UInt16([u16; 2], Option<u16>),
-    UInt32([u32; 2], Option<u32>),
-    UInt64([u64; 2], Option<u64>),
-    Float32([f32; 2], Option<f32>),
-    Float64([f64; 2], Option<f64>),
-    StringAscii,
-}
-
-#[macro_export]
-macro_rules! dimension_constraints_go {
-    ($expr:expr, $DT:ident, $range:pat, $extent:pat, $then:expr, $string:expr) => {{
-        dimension_constraints_go!(
-            $expr, $DT, $range, $extent, $then, $then, $string
-        )
-    }};
-    ($expr:expr, $DT:ident, $range:pat, $extent:pat, $integral:expr, $float:expr, $string:expr) => {{
-        use $crate::array::dimension::DimensionConstraints;
-        match $expr {
-            #[allow(unused_variables)]
-            DimensionConstraints::Int8($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = i8;
-                $integral
-            }
-            #[allow(unused_variables)]
-            DimensionConstraints::Int16($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = i16;
-                $integral
-            }
-            #[allow(unused_variables)]
-            DimensionConstraints::Int32($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = i32;
-                $integral
-            }
-            #[allow(unused_variables)]
-            DimensionConstraints::Int64($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = i64;
-                $integral
-            }
-            #[allow(unused_variables)]
-            DimensionConstraints::UInt8($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = u8;
-                $integral
-            }
-            #[allow(unused_variables)]
-            DimensionConstraints::UInt16($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = u16;
-                $integral
-            }
-            #[allow(unused_variables)]
-            DimensionConstraints::UInt32($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = u32;
-                $integral
-            }
-            #[allow(unused_variables)]
-            DimensionConstraints::UInt64($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = u64;
-                $integral
-            }
-            #[allow(unused_variables)]
-            DimensionConstraints::Float32($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = f32;
-                $float
-            }
-            #[allow(unused_variables)]
-            DimensionConstraints::Float64($range, $extent) => {
-                #[allow(dead_code)]
-                type $DT = f64;
-                $float
-            }
-            DimensionConstraints::StringAscii => $string,
-        }
-    }};
-}
-
-macro_rules! dimension_constraints_impl {
-    ($($V:ident : $U:ty),+) => {
-        $(
-            impl From<[$U; 2]> for DimensionConstraints {
-                fn from(value: [$U; 2]) -> DimensionConstraints {
-                    DimensionConstraints::$V(value, None)
-                }
-            }
-
-            impl From<&[$U; 2]> for DimensionConstraints {
-                fn from(value: &[$U; 2]) -> DimensionConstraints {
-                    DimensionConstraints::$V([value[0], value[1]], None)
-                }
-            }
-
-            impl From<([$U; 2], $U)> for DimensionConstraints {
-                fn from(value: ([$U; 2], $U)) -> DimensionConstraints {
-                    DimensionConstraints::$V([value.0[0], value.0[1]], Some(value.1))
-                }
-            }
-
-            impl From<(&[$U; 2], $U)> for DimensionConstraints {
-                fn from(value: (&[$U; 2], $U)) -> DimensionConstraints {
-                    DimensionConstraints::$V([value.0[0], value.0[1]], Some(value.1))
-                }
-            }
-
-            impl From<([$U; 2], Option<$U>)> for DimensionConstraints {
-                fn from(value: ([$U; 2], Option<$U>)) -> DimensionConstraints {
-                    DimensionConstraints::$V([value.0[0], value.0[1]], value.1)
-                }
-            }
-
-            impl From<(&[$U; 2], Option<$U>)> for DimensionConstraints {
-                fn from(value: (&[$U; 2], Option<$U>)) -> DimensionConstraints {
-                    DimensionConstraints::$V([value.0[0], value.0[1]], value.1)
-                }
-            }
-        )+
-    }
-}
-
-dimension_constraints_impl!(Int8: i8, Int16: i16, Int32: i32, Int64: i64);
-dimension_constraints_impl!(UInt8: u8, UInt16: u16, UInt32: u32, UInt64: u64);
-dimension_constraints_impl!(Float32: f32, Float64: f64);
-
-impl DimensionConstraints {
-    /// Returns a [Datatype] which represents the physical type of this constraint.
-    pub fn physical_datatype(&self) -> Datatype {
-        match self {
-            Self::UInt8(_, _) => Datatype::UInt8,
-            Self::UInt16(_, _) => Datatype::UInt16,
-            Self::UInt32(_, _) => Datatype::UInt32,
-            Self::UInt64(_, _) => Datatype::UInt64,
-            Self::Int8(_, _) => Datatype::Int8,
-            Self::Int16(_, _) => Datatype::Int16,
-            Self::Int32(_, _) => Datatype::Int32,
-            Self::Int64(_, _) => Datatype::Int64,
-            Self::Float32(_, _) => Datatype::Float32,
-            Self::Float64(_, _) => Datatype::Float64,
-            Self::StringAscii => Datatype::StringAscii,
-        }
-    }
-
-    pub fn cell_val_num(&self) -> CellValNum {
-        match self {
-            DimensionConstraints::StringAscii => CellValNum::Var,
-            _ => CellValNum::single(),
-        }
-    }
-
-    pub fn verify_type_compatible(
-        &self,
-        datatype: Datatype,
-    ) -> TileDBResult<()> {
-        dimension_constraints_go!(
-            self,
-            DT,
-            _range,
-            _extent,
-            {
-                if !datatype.is_compatible_type::<DT>() {
-                    return Err(Error::Datatype(
-                        DatatypeErrorKind::TypeMismatch {
-                            user_type: std::any::type_name::<DT>().to_owned(),
-                            tiledb_type: datatype,
-                        },
-                    ));
-                }
-            },
-            {
-                if !matches!(datatype, Datatype::StringAscii) {
-                    return Err(Error::Datatype(
-                        DatatypeErrorKind::InvalidDatatype {
-                            context: Some(
-                                "DimensionConstraints::StringAscii".to_owned(),
-                            ),
-                            found: datatype,
-                            expected: Datatype::StringAscii,
-                        },
-                    ));
-                }
-            }
-        );
-
-        Ok(())
-    }
-
-    pub(crate) fn domain_ptr(&self) -> *const std::ffi::c_void {
-        dimension_constraints_go!(
-            self,
-            DT,
-            range,
-            _extent,
-            range.as_ptr() as *const DT as *const std::ffi::c_void,
-            std::ptr::null()
-        )
-    }
-
-    pub(crate) fn extent_ptr(&self) -> *const std::ffi::c_void {
-        dimension_constraints_go!(
-            self,
-            DT,
-            _range,
-            extent,
-            {
-                if let Some(extent) = extent {
-                    extent as *const DT as *const std::ffi::c_void
-                } else {
-                    std::ptr::null()
-                }
-            },
-            std::ptr::null()
-        )
-    }
-
-    /// Returns the number of cells spanned by this constraint, if applicable
-    pub fn num_cells(&self) -> Option<u128> {
-        let (low, high) = crate::dimension_constraints_go!(
-            self,
-            _DT,
-            [low, high],
-            _,
-            (i128::from(*low), i128::from(*high)),
-            return None,
-            return None
-        );
-
-        Some(1 + (high - low) as u128)
-    }
-
-    /// Returns the number of cells spanned by a
-    /// single tile under this constraint, if applicable
-    pub fn num_cells_per_tile(&self) -> Option<usize> {
-        crate::dimension_constraints_go!(
-            self,
-            _DT,
-            _,
-            extent,
-            extent.map(|extent| {
-                #[allow(clippy::unnecessary_fallible_conversions)]
-                // this `unwrap` should be safe, validation will confirm nonzero
-                usize::try_from(extent).unwrap()
-            }),
-            None,
-            None
-        )
-    }
-
-    /// Returns the domain of the dimension constraint, if present, as a range.
-    pub fn domain(&self) -> Option<SingleValueRange> {
-        crate::dimension_constraints_go!(
-            self,
-            _DT,
-            [low, high],
-            _,
-            Some(SingleValueRange::from(&[*low, *high])),
-            None
-        )
     }
 }
 
@@ -521,210 +236,8 @@ impl From<Builder> for Dimension {
     }
 }
 
-/// Encapsulation of data needed to construct a Dimension
-#[derive(Clone, Debug, Deserialize, OptionSubset, PartialEq, Serialize)]
-pub struct DimensionData {
-    pub name: String,
-    pub datatype: Datatype,
-    pub constraints: DimensionConstraints,
-
-    /// Optional filters to apply to the dimension. If None or Some(empty),
-    /// then filters will be inherited from the schema's `coordinate_filters`
-    /// field when the array is constructed.
-    pub filters: Option<FilterListData>,
-}
-
-impl DimensionData {
-    pub fn cell_val_num(&self) -> CellValNum {
-        self.constraints.cell_val_num()
-    }
-}
-
-#[cfg(any(test, feature = "proptest-strategies"))]
-impl DimensionData {
-    /// Returns a strategy for generating values of this dimension's type
-    /// which fall within the domain of this dimension.
-    pub fn value_strategy(&self) -> crate::query::strategy::FieldValueStrategy {
-        use crate::query::strategy::FieldValueStrategy;
-        use proptest::prelude::*;
-
-        dimension_constraints_go!(
-            self.constraints,
-            DT,
-            ref domain,
-            _,
-            FieldValueStrategy::from((domain[0]..=domain[1]).boxed()),
-            {
-                assert_eq!(self.datatype, Datatype::StringAscii);
-                FieldValueStrategy::from(any::<u8>().boxed())
-            }
-        )
-    }
-
-    /// Returns a strategy for generating subarray ranges which fall within
-    /// the domain of this dimension.
-    ///
-    /// `cell_bound` is an optional restriction on the number of possible values
-    /// which the strategy is allowed to return.
-    ///
-    /// If `cell_bound` is `None`, then this function always returns `Some`.
-    pub fn subarray_strategy(
-        &self,
-        cell_bound: Option<usize>,
-    ) -> Option<proptest::strategy::BoxedStrategy<crate::range::Range>> {
-        use proptest::prelude::Just;
-        use proptest::strategy::Strategy;
-
-        use crate::range::{Range, VarValueRange};
-
-        dimension_constraints_go!(
-            self.constraints,
-            DT,
-            ref domain,
-            _,
-            {
-                let cell_bound = cell_bound
-                    .map(|bound| DT::try_from(bound).unwrap_or(DT::MAX))
-                    .unwrap_or(DT::MAX);
-
-                let domain_lower = domain[0];
-                let domain_upper = domain[1];
-                let strat =
-                    (domain_lower..=domain_upper).prop_flat_map(move |lb| {
-                        let ub = std::cmp::min(
-                            domain_upper,
-                            lb.checked_add(cell_bound).unwrap_or(DT::MAX),
-                        );
-                        (Just(lb), lb..=ub).prop_map(|(min, max)| {
-                            Range::Single(SingleValueRange::from(&[min, max]))
-                        })
-                    });
-                Some(strat.boxed())
-            },
-            {
-                if cell_bound.is_some() {
-                    /*
-                     * This can be implemented, but there's some ambiguity about
-                     * what it should mean when precision goes out the window,
-                     * so wait until there's a use case to decide.
-                     */
-                    return None;
-                }
-
-                let domain_lower = domain[0];
-                let domain_upper = domain[1];
-                let strat =
-                    (domain_lower..=domain_upper).prop_flat_map(move |lb| {
-                        (Just(lb), (lb..=domain_upper)).prop_map(
-                            |(min, max)| {
-                                Range::Single(SingleValueRange::from(&[
-                                    min, max,
-                                ]))
-                            },
-                        )
-                    });
-                Some(strat.boxed())
-            },
-            {
-                // DimensionConstraints::StringAscii
-                let strat_bound =
-                    proptest::string::string_regex("[ -~]*").unwrap().boxed();
-
-                if cell_bound.is_some() {
-                    /*
-                     * This is not tractible unless there is a bound on the string length.
-                     * There isn't one since `StringAscii` is only allowed as a dimension
-                     * type in sparse arrays.
-                     */
-                    return None;
-                }
-
-                let strat = (strat_bound.clone(), strat_bound).prop_map(
-                    |(ascii1, ascii2)| {
-                        let (lb, ub) = if ascii1 < ascii2 {
-                            (ascii1, ascii2)
-                        } else {
-                            (ascii2, ascii1)
-                        };
-                        Range::Var(VarValueRange::from((lb, ub)))
-                    },
-                );
-                Some(strat.boxed())
-            }
-        )
-    }
-}
-
-impl Display for DimensionData {
-    fn fmt(&self, f: &mut Formatter) -> FmtResult {
-        write!(f, "{}", json!(*self))
-    }
-}
-
-impl TryFrom<&Dimension> for DimensionData {
-    type Error = crate::error::Error;
-
-    fn try_from(dim: &Dimension) -> TileDBResult<Self> {
-        let datatype = dim.datatype()?;
-        let constraints = physical_type_go!(datatype, DT, {
-            let domain = dim.domain::<DT>()?;
-            let extent = dim.extent::<DT>()?;
-            if let Some(domain) = domain {
-                DimensionConstraints::from((domain, extent))
-            } else {
-                assert!(extent.is_none());
-                DimensionConstraints::StringAscii
-            }
-        });
-
-        Ok(DimensionData {
-            name: dim.name()?,
-            datatype,
-            constraints,
-            filters: {
-                let fl = FilterListData::try_from(&dim.filters()?)?;
-                if fl.is_empty() {
-                    None
-                } else {
-                    Some(fl)
-                }
-            },
-        })
-    }
-}
-
-impl TryFrom<Dimension> for DimensionData {
-    type Error = crate::error::Error;
-
-    fn try_from(dim: Dimension) -> TileDBResult<Self> {
-        Self::try_from(&dim)
-    }
-}
-
-impl Factory for DimensionData {
-    type Item = Dimension;
-
-    fn create(&self, context: &Context) -> TileDBResult<Self::Item> {
-        let mut b = Builder::new(
-            context,
-            &self.name,
-            self.datatype,
-            self.constraints.clone(),
-        )?;
-
-        if let Some(fl) = self.filters.as_ref() {
-            b = b.filters(fl.create(context)?)?;
-        }
-
-        Ok(b.cell_val_num(self.cell_val_num())?.build())
-    }
-}
-
 #[cfg(feature = "arrow")]
 pub mod arrow;
-
-#[cfg(any(test, feature = "proptest-strategies"))]
-pub mod strategy;
 
 #[cfg(test)]
 mod tests {
@@ -968,93 +481,5 @@ mod tests {
             assert_eq!(cmp, cmp);
             assert_ne!(base, cmp);
         }
-    }
-
-    #[test]
-    fn subarray_strategy_dense() {
-        use super::strategy::Requirements;
-        use crate::array::ArrayType;
-        use crate::range::{Range, SingleValueRange};
-        use proptest::prelude::*;
-        use proptest::strategy::Strategy;
-        use std::rc::Rc;
-
-        let req = Requirements {
-            array_type: Some(ArrayType::Dense),
-            ..Default::default()
-        };
-        let strat = (
-            any_with::<DimensionData>(req),
-            prop_oneof![Just(None), any::<usize>().prop_map(Some)],
-        )
-            .prop_flat_map(|(d, cell_bound)| {
-                let subarray_strat = d
-                    .subarray_strategy(cell_bound)
-                    .expect("Dense dimension must have a subarray strategy");
-                (Just(Rc::new(d)), Just(cell_bound), subarray_strat)
-            });
-
-        proptest!(|((d, cell_bound, s) in strat)| {
-            if let Some(bound) = cell_bound {
-                assert!(s.num_cells().unwrap() <= bound as u128);
-            }
-            if let Some(num_cells) = d.constraints.num_cells() {
-                assert!(s.num_cells().unwrap() <= num_cells);
-            }
-            let Range::Single(s) = s else {
-                unreachable!("Unexpected range for dense dimension: {:?}", s)
-            };
-            match s {
-                SingleValueRange::Int8(start, end) => {
-                    let DimensionConstraints::Int8([lb, ub], _) = d.constraints else { unreachable!() };
-                    assert!(lb <= start);
-                    assert!(end <= ub);
-                    assert_eq!(Some((end - start + 1) as u128), s.num_cells());
-                }
-                SingleValueRange::Int16(start, end) => {
-                    let DimensionConstraints::Int16([lb, ub], _) = d.constraints else { unreachable!() };
-                    assert!(lb <= start);
-                    assert!(end <= ub);
-                    assert_eq!(Some((end - start + 1) as u128), s.num_cells());
-                }
-                SingleValueRange::Int32(start, end) => {
-                    let DimensionConstraints::Int32([lb, ub], _) = d.constraints else { unreachable!() };
-                    assert!(lb <= start);
-                    assert!(end <= ub);
-                    assert_eq!(Some((end - start + 1) as u128), s.num_cells());
-                }
-                SingleValueRange::Int64(start, end) => {
-                    let DimensionConstraints::Int64([lb, ub], _) = d.constraints else { unreachable!() };
-                    assert!(lb <= start);
-                    assert!(end <= ub);
-                    assert_eq!(Some((end - start + 1) as u128), s.num_cells());
-                }
-                SingleValueRange::UInt8(start, end) => {
-                    let DimensionConstraints::UInt8([lb, ub], _) = d.constraints else { unreachable!() };
-                    assert!(lb <= start);
-                    assert!(end <= ub);
-                    assert_eq!(Some((end - start + 1) as u128), s.num_cells());
-                }
-                SingleValueRange::UInt16(start, end) => {
-                    let DimensionConstraints::UInt16([lb, ub], _) = d.constraints else { unreachable!() };
-                    assert!(lb <= start);
-                    assert!(end <= ub);
-                    assert_eq!(Some((end - start + 1) as u128), s.num_cells());
-                }
-                SingleValueRange::UInt32(start, end) => {
-                    let DimensionConstraints::UInt32([lb, ub], _) = d.constraints else { unreachable!() };
-                    assert!(lb <= start);
-                    assert!(end <= ub);
-                    assert_eq!(Some((end - start + 1) as u128), s.num_cells());
-                }
-                SingleValueRange::UInt64(start, end) => {
-                    let DimensionConstraints::UInt64([lb, ub], _) = d.constraints else { unreachable!() };
-                    assert!(lb <= start);
-                    assert!(end <= ub);
-                    assert_eq!(Some((end - start + 1) as u128), s.num_cells());
-                },
-                s => unreachable!("Unexpected range type for dense dimension: {:?}", s)
-            }
-        });
     }
 }
