+++ tiledb/api/src/lib.rs
@@ -1,9 +1,5 @@
 extern crate anyhow;
-extern crate serde;
-extern crate serde_json;
 extern crate thiserror;
-#[macro_use]
-extern crate tiledb_proc_macro;
 extern crate tiledb_sys as ffi;
 extern crate tiledb_utils as util;
 
@@ -40,6 +36,9 @@ macro_rules! out_ptr {
     };
 }
 
+pub use tiledb_common::physical_type_go;
+pub use tiledb_common::range;
+
 pub mod array;
 pub mod config;
 pub mod context;
@@ -51,8 +50,6 @@ pub mod group;
 pub mod key;
 pub mod metadata;
 pub mod query;
-#[macro_use]
-pub mod range;
 pub mod stats;
 pub mod string;
 pub mod vfs;
@@ -79,33 +76,3 @@ pub use array::Array;
 pub use context::{Context, ContextBound};
 pub use datatype::Datatype;
 pub type Result<T> = std::result::Result<T, error::Error>;
-
-pub trait Factory {
-    type Item;
-
-    fn create(&self, context: &context::Context) -> Result<Self::Item>;
-}
-
-mod private {
-    // The "sealed trait" pattern is a way to prevent downstream crates from implementing traits
-    // that you don't think they should implement. If you have `trait Foo: Sealed`, then
-    // downstream crates cannot `impl Foo` because they cannot `impl Sealed`.
-    //
-    // Semantic versioning is one reason you might want this.
-    // We currently use this as a bound for `datatype::PhysicalType` and `datatype::LogicalType`
-    // so that we won't accept something that we don't know about for the C API calls.
-    pub trait Sealed {}
-
-    macro_rules! sealed {
-        ($($DT:ty),+) => {
-            $(
-                impl crate::private::Sealed for $DT {}
-            )+
-        }
-    }
-
-    pub(crate) use sealed;
-}
-
-#[cfg(any(test, feature = "proptest-strategies"))]
-pub mod strategy;
