+++ tiledb/common/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "tiledb-common"
+edition.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+arrow-schema = { workspace = true, optional = true }
+paste = { workspace = true }
+proptest = { workspace = true, optional = true }
+serde = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
+thiserror = { workspace = true }
+tiledb-proc-macro = { workspace = true, optional = true }
+tiledb-sys = { workspace = true }
+tiledb-utils = { workspace = true, optional = true }
+
+[dev-dependencies]
+anyhow = { workspace = true }
+proptest = { workspace = true }
+
+[features]
+default = []
+arrow = ["dep:arrow-schema"]
+option-subset = ["dep:tiledb-proc-macro", "dep:tiledb-utils"]
+proptest-strategies = ["dep:proptest"]
+serde = ["dep:serde", "dep:serde_json"]
