+++ tiledb/proptest-integrations/src/array/attribute.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod tests {}
