+++ tiledb/proptest-integrations/src/lib.rs
@@ -1,0 +1,1 @@
+mod array;
