+++ tiledb/queries/Cargo.toml
@@ -5,7 +5,7 @@ rust-version.workspace = true
 version.workspace = true
 
 [dependencies]
-tiledb = { workspace = true }
+tiledb-api = { workspace = true }
 
 [build-dependencies]
 pkg-config = { workspace = true }
