+++ tiledb/serde/src/array/domain/strategy.rs
@@ -221,7 +221,6 @@ mod tests {
         });
     }
 
-    #[ignore = "takes a long time due to shrink iters on name, we should do custom shrink strategy"]
     #[test]
     fn domain_shrinking() {
         let strat = any::<DomainData>();
