+++ tiledb/serde/src/array/enumeration/strategy.rs
