+++ tiledb/serde/src/array/mod.rs
@@ -1,0 +1,4 @@
+pub mod attribute;
+pub mod dimension;
+pub mod domain;
+pub mod schema;
