+++ tiledb/sys/src/datatype.rs
@@ -11,6 +11,4 @@ extern "C" {
         str_: *const ::std::os::raw::c_char,
         datatype: *mut tiledb_datatype_t,
     ) -> capi_return_t;
-
-    pub fn tiledb_datatype_size(type_: tiledb_datatype_t) -> u64;
 }
